+++ c_lib/lexer.c
@@ -2,7 +2,6 @@
 #include <ctype.h>
 #include <string.h>
 #include <sys/types.h>
-#include <sys/uio.h>
 #include <unistd.h>
 #include <fcntl.h>